+++ src/manifest.rs
@@ -129,7 +129,7 @@ pub(crate) fn with(cx: &Context, f: impl FnOnce() -> Result<()>) -> Result<()> {
                         "--no-private is not supported yet with workspace with private root crate"
                     );
                 }
-                private_crates.push(manifest_path);
+                private_crates.push(manifest_path.canonicalize()?);
             } else if is_root && no_private {
                 //
             } else if no_dev_deps {
@@ -197,7 +197,7 @@ fn remove_dev_deps(doc: &mut toml_edit::Document) {
 fn remove_private_crates(
     doc: &mut toml_edit::Document,
     metadata: &metadata::Metadata,
-    private_crates: &[&PathBuf],
+    private_crates: &[PathBuf],
 ) -> Result<()> {
     let table = doc.as_table_mut();
     if let Some(workspace) = table.get_mut("workspace").and_then(toml_edit::Item::as_table_like_mut)
@@ -209,7 +209,7 @@ fn remove_private_crates(
                 if let Some(member) = members.get(i).and_then(toml_edit::Value::as_str) {
                     let manifest_path =
                         metadata.workspace_root.join(member).join("Cargo.toml").canonicalize()?;
-                    if private_crates.iter().any(|p| **p == manifest_path) {
+                    if private_crates.iter().any(|p| *p == manifest_path) {
                         members.remove(i);
                         continue;
                     }